% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3Dtree.R
\name{Plot3DTree}
\alias{Plot3DTree}
\title{Main function to plot tree}
\usage{
Plot3DTree(phy, width.scaling = 1, new.window = TRUE)
}
\arguments{
\item{phy}{Phylo object}

\item{width.scaling}{How wide to make each branch relative to program defaults}

\item{new.window}{Should we open a new plotting window?}
}
\description{
Will create a 3D tree. It positions tips randomly
}
\examples{
phy <- ape::rcoal(10)
}
